Bugfix/installer script

+++ installer/tools/installer.ps1
@@ -1,7 +1,7 @@
 param(
     [string]$version="v23.08.02-test1",
     [string]$msiUrl ,
-    [string]$installPath = $env:APPDATA,
+    [string]$installPath = "$env:UserProfile\AppData\local\Sensing-Dev",
     [string]$relativeScriptPath = "tools\Env.ps1"
 )
 # Check if the MSI URL and Install Path are provided
@@ -19,7 +19,10 @@ if (-not $msiUrl ) {
     if ($version.StartsWith("v")) {
         $version = $version.Substring(1)
     }
-    $msiUrl = "https://github.com/Sensing-Dev/sensing-dev-installer/releases/download/v${version}/sensing-dev-installer-${version}-win64.msi"
+    # if the tag has "-"
+    $majorVersion = $version -split "-"
+    $majorVersion = $majorVersion[0]
+    $msiUrl = "https://github.com/Sensing-Dev/sensing-dev-installer/releases/download/v${version}/sensing-dev-installer-${majorVersion}-win64.msi"
 }
 
 # Download MSI to a temp location
@@ -28,7 +31,9 @@ Invoke-WebRequest -Uri $msiUrl -OutFile $tempMsiPath -Verbose
 
 # Install MSI to specified path
 # Note: This assumes the MSI accepts TARGETDIR as an argument for installation directory. Some MSIs might not.
-Start-Process -Wait -FilePath "msiexec.exe" -ArgumentList "/i `"$tempMsiPath`" TARGETDIR=`"$installPath`" /qn"
+
+Start-Process -Wait -FilePath "msiexec.exe" -ArgumentList "/i `"$tempMsiPath`" INSTALL_ROOT=`"$installPath`" /qn"
+
 
 # Run .ps1 file from the installed package
 $ps1ScriptPath = Join-Path -Path $installPath -ChildPath $relativeScriptPath
@@ -39,4 +44,6 @@ if (Test-Path -Path $ps1ScriptPath -PathType Leaf) {
 }
 
 # Cleanup
-Remove-Item -Path $tempMsiPath -Force
+if (Test-Path -Path $tempMsiPath ){
+    Remove-Item -Path $tempMsiPath -Force
+}
